# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Youtube Download REST",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/api/download": {
            "get": {
                "description": "Download video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Download video",
                "operationId": "download",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL of youtube video",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Itag of youtube video",
                        "name": "itag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/api/info": {
            "post": {
                "description": "Get information of video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get video info",
                "operationId": "get-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL of youtube video",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}